+++ shaders/common/setting_defines.glsl
@@ -554,6 +554,7 @@ const float sunPathRotation = -30.0; // [-80.0 -75.0 -70.0 -65.0 -60.0 -55.0 -50
 
 
 
+#define DO_COLOR_CODED_GBUFFERS 0 // [0 1]
 #define HAND_DEPTH 0.56 // [ 0.3 0.32 0.34 0.36 0.38 0.4 0.42 0.44 0.46 0.48 0.5 0.52 0.54 0.56 0.58 0.6 0.62 0.64 0.66 0.68 0.7 0.72 0.74 0.76 0.78 0.8]
 #define SHOW_DANGEROUS_LIGHT 0 // [0 1]
 #define CLOUD_COMPATIBILITY 0 // [0 1]