+++ shaders/program/gbuffers_armor_glint.glsl
@@ -10,6 +10,9 @@ void main() {
 	vec4 color = texture2D(MAIN_TEXTURE, texcoord) * glcolor;
 	
 	/* DRAWBUFFERS:0 */
+	#if DO_COLOR_CODED_GBUFFERS == 1
+		color = vec4(1.0, 0.0, 0.5, 1.0);
+	#endif
 	color.rgb *= 0.5;
 	gl_FragData[0] = vec4(color);