+++ shaders/program/gbuffers_basic.glsl
@@ -11,6 +11,9 @@ void main() {
 	if (color.a < 0.01) discard;
 	
 	/* DRAWBUFFERS:02 */
+	#if DO_COLOR_CODED_GBUFFERS == 1
+		color = vec4(0.0, 0.0, 0.0, 1.0);
+	#endif
 	color.rgb *= 0.5;
 	gl_FragData[0] = vec4(color);
 	gl_FragData[1] = vec4(