+++ shaders/program/gbuffers_particles_translucent.glsl
@@ -49,6 +49,9 @@ void main() {
 	
 	
 	/* DRAWBUFFERS:03 */
+	#if DO_COLOR_CODED_GBUFFERS == 1
+		color = vec4(0.5, 0.75, 0.75, 1.0);
+	#endif
 	color.rgb *= 0.5;
 	gl_FragData[0] = vec4(color);
 	gl_FragData[1] = vec4(