+++ shaders/program/gbuffers_skybasic.glsl
@@ -6,6 +6,9 @@ in_out vec3 color;
 
 void main() {
 	/* DRAWBUFFERS:5 */
+	#if DO_COLOR_CODED_GBUFFERS == 1
+		vec3 color = vec3(1.0, 1.0, 0.5);
+	#endif
 	gl_FragData[0] = vec4(color, 1.0);
 }