+++ shaders/program/gbuffers_water.glsl
@@ -226,6 +226,9 @@ void main() {
 	
 	
 	/* DRAWBUFFERS:03 */
+	#if DO_COLOR_CODED_GBUFFERS == 1
+		color = vec4(0.0, 0.0, 1.0, 1.0);
+	#endif
 	color.rgb *= 0.5;
 	gl_FragData[0] = color;
 	gl_FragData[1] = vec4(